+++ test/mail_test.py
@@ -1,3 +1,5 @@
+import pytest
+
 import src.mail
 
 """
@@ -5,6 +7,8 @@ Test by sending mail to self
 """
 
 
+
+@pytest.mark.skip(reason="Shouldn't be run automatically")
 def test_send_mail():
     subject = "Subject abcd"
     content = "Some text"
